HDF5 linker support
This temporarily adds explicit flags to the HDF5 library.

Default HDF5 on Gaea is 1.14.3, which contains an error that raises
floating point exceptions, and preventing their usage in debug tests.

There is also a module issue on Gaea preventing us from changing the
default version of HDF5, forcing us to use 1.14.3.

This patch gets around this issue by explcitly providing flags to the
selected version of HDF5.  If HDF5 1.12.x is loaded, then this should
override the default setting.

This should only be needed until the modules on Gaea are fixed.