Disable FPEs in DEBUG due to HDF5
This patch disables floating point exceptions in the DEBUG builds.

Default HDF5 on Gaea is 1.14.3, which contains an error that raises
floating point exceptions, and preventing their usage in debug tests.

There is also a module issue on Gaea preventing us from changing the
default version of HDF5, forcing us to use 1.14.3.

There seems to be no viable way out of this problem, since 1.14.3 has
been forced upon netCDF.  So for now we have to disable FPEs.

Hopefully this is a temporary solution.

The contents of ncrc5-gnu.mk and ncrc5-gcc.mk have also been
consolidated, and ncrc5-gnu.mk is now a symlink to the other.